Extract XcTest lib finder into fixture
